rename endpoints to reflect deduplicates specific actions
namely:
  check_entities -> check_human_duplicates
  check_entity.js -> check_human_duplicate
  collect_entities -> collect_human_duplicates
