+++ server/controllers/tasks/collect_human_duplicates.ts
@@ -6,7 +6,7 @@ import { initJobQueue } from '#db/level/jobs'
 import { waitForCPUsLoadToBeBelow } from '#lib/os'
 import { success, info, logError, LogError } from '#lib/utils/logs'
 import config from '#server/config'
-import checkEntity from './lib/check_entity.js'
+import checkHumanDuplicate from './lib/check_human_duplicate.js'
 
 const { nice } = config
 
@@ -34,7 +34,7 @@ function addEntitiesToQueueSequentially (refresh) {
     const uris = await getNextInvHumanUrisBatch(pagination)
     pagination.total += uris.length
     if (uris.length === 0) {
-      success(pagination.total, 'done. total entities queued:')
+      success(pagination.total, 'Done. Total entities queued:')
     } else {
       const filteredUris = await getFilteredUris(uris, refresh)
       await invTasksEntitiesQueue.pushBatch(filteredUris)
@@ -69,7 +69,7 @@ async function deduplicateWorker (jobId, uri) {
     // to give the priority to more urgent matters,
     // such as answering users requests
     if (nice) await waitForCPUsLoadToBeBelow({ threshold: 0.5 })
-    await checkEntity(uri)
+    await checkHumanDuplicate(uri)
   } catch (err) {
     // Prevent crashing the queue for non-critical errors
     // Example of 400 error: the entity has already been redirected
